requirements.txt
babel
bottle
irc
python-dotenv
pytz
requests